requirements.txt
djangorestframework==3.7.7
virtualenvwrapper==4.8.2
django-flat-theme==1.1.4
